integrate unit test setup into test class
